requirements.txt
appdirs==1.4.4
appnope==0.1.2
attrs==20.3.0
backcall==0.2.0
boto3==1.16.63
botocore==1.19.63
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cliff==3.6.0
cloudbridge==2.1.0
cmd2==1.5.0
colorama==0.4.4
cryptography==3.4.1
debtcollector==2.2.0
decorator==4.4.2
deprecation==2.1.0
dogpile.cache==1.1.2
google-api-core==1.25.1
google-api-python-client==1.12.8
google-auth==1.25.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.19.0
idna==2.10
importlib-metadata==3.4.0
ipython==7.20.0
ipython-genutils==0.2.0
iso8601==0.1.14
jedi==0.18.0
jmespath==0.10.0
jsonpatch==1.28
jsonpointer==2.0
keystoneauth1==4.3.0
msgpack==1.0.2
munch==2.5.0
netaddr==0.8.0
netifaces==0.10.9
oauth2client==4.1.3
openstacksdk==0.52.0
os-client-config==2.1.0
os-service-types==1.7.0
osc-lib==2.3.1
oslo.config==8.4.0
oslo.context==3.1.1
oslo.i18n==5.0.1
oslo.log==4.4.0
oslo.serialization==4.1.0
oslo.utils==4.7.0
packaging==20.9
parso==0.8.1
pbr==5.5.1
pexpect==4.8.0
pickleshare==0.7.5
prettytable==0.7.2
prompt-toolkit==3.0.14
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyeventsystem==0.1.0
Pygments==2.7.4
pyparsing==2.4.7
pyperclip==1.8.1
python-dateutil==2.8.1
python-keystoneclient==4.1.1
python-neutronclient==7.2.1
python-novaclient==17.2.1
python-swiftclient==3.10.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requestsexceptions==1.4.0
rfc3986==1.4.0
rsa==4.7
s3transfer==0.3.4
semantic-version==2.8.5
setuptools-rust==0.11.6
simplejson==3.17.2
six==1.15.0
stevedore==3.3.0
tenacity==6.3.1
toml==0.10.2
traitlets==5.0.5
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.4.0
